competences/requirements.txt
Flask==2.0.2
pymongo==3.12.0
requests==2.26.0
python-consul==1.1.0